PrintBeacon - fix bug
I think it fixed but not sure. I change 0 to 255 as I read again the comments of the fanction and understand it needed to be 255 instead of 0 as the supervisor is on the same board.
need another check on the sat (I did one but not sure the result is fine. It's just not an error.)