create commands to the trxvu
create 2 commands to the trxvu. Get and Set Beacon interval.
change the ACK from "ACK_UPDATE_BEACON_TIME_DELAY" to "ACK_UPDATE_BEACON_INTERVAL".
and check the trxvu_command_router in the CommandDictionary to the two command I created and need check to see if it's working and to check all the error_msg acks.